third_party/blink/web_tests/platform/linux-chrome/external/wpt/clipboard-apis/async-custom-formats-write-read-without-web-prefix.tentative.https-expected.txt

This is a testharness.js-based test.
[FAIL] navigator.clipboard.write() for custom format with web prefix, but Blob type without web prefix
  assert_equals: expected "input data 1" but got ""
Harness: the test ran to completion.
